connect/data/load/requirements.txt
black
pylint
ipdb
pytest
typer[all]
psycopg2-binary